feat: move sign al sign transaction methods to vechain-base-signer tests
