+++ packages/network/tests/thor-client/transactions/fixture.ts
@@ -8,14 +8,8 @@ import { BUILT_IN_CONTRACTS } from '../../built-in-fixture';
 import {
     TEST_ACCOUNTS,
     TESTING_CONTRACT_ABI,
-    TESTING_CONTRACT_ADDRESS,
-    TESTNET_DELEGATE_URL
+    TESTING_CONTRACT_ADDRESS
 } from '../../fixture';
-import {
-    InvalidSecp256k1PrivateKeyError,
-    TransactionDelegationError
-} from '@vechain/sdk-errors';
-import { type SignTransactionOptions } from '../../../src';
 
 /**
  * Some random transaction nonces to use into tests
@@ -298,130 +292,6 @@ const buildTransactionBodyClausesTestCases = [
     }
 ];
 
-/**
- * signTransaction test cases
- * Has both correct and incorrect for solo and an example of using delegatorUrl on testnet
- */
-const signTransactionTestCases = {
-    solo: {
-        correct: [
-            {
-                description: 'Should sign a transaction without delegation',
-                origin: TEST_ACCOUNTS.TRANSACTION.TRANSACTION_SENDER,
-                isDelegated: false,
-                expected: {
-                    body: {
-                        chainTag: 246,
-                        clauses: [
-                            {
-                                data: '0xb6b55f25000000000000000000000000000000000000000000000000000000000000007b',
-                                to: '0xb2c20a6de401003a671659b10629eb82ff254fb8',
-                                value: 0
-                            }
-                        ],
-                        dependsOn: null,
-                        expiration: 32,
-                        gas: 57491,
-                        gasPriceCoef: 0,
-                        reserved: undefined
-                    }
-                }
-            },
-            {
-                description:
-                    'Should sign a transaction with private key delegation',
-                origin: TEST_ACCOUNTS.TRANSACTION.TRANSACTION_SENDER,
-                options: {
-                    delegatorPrivateKey:
-                        TEST_ACCOUNTS.TRANSACTION.DELEGATOR.privateKey
-                } satisfies SignTransactionOptions,
-                isDelegated: true,
-                expected: {
-                    body: {
-                        chainTag: 246,
-                        clauses: [
-                            {
-                                data: '0xb6b55f25000000000000000000000000000000000000000000000000000000000000007b',
-                                to: '0xb2c20a6de401003a671659b10629eb82ff254fb8',
-                                value: 0
-                            }
-                        ],
-                        dependsOn: null,
-                        expiration: 32,
-                        gas: 57491,
-                        gasPriceCoef: 0,
-                        reserved: {
-                            features: 1
-                        }
-                    }
-                }
-            }
-        ],
-        incorrect: [
-            {
-                description:
-                    "Should throw error when delegator's private key is invalid",
-                origin: TEST_ACCOUNTS.TRANSACTION.TRANSACTION_SENDER,
-                options: {
-                    delegatorPrivateKey: '0x'
-                } satisfies SignTransactionOptions,
-                isDelegated: true,
-                expectedError: InvalidSecp256k1PrivateKeyError
-            },
-            {
-                description:
-                    'Should throw error when delegator private key is invalid',
-                origin: {
-                    privateKey: '0x'
-                },
-                isDelegated: true,
-                expectedError: InvalidSecp256k1PrivateKeyError
-            },
-            {
-                description:
-                    "Should throw error when using delegator url on solo network due to no server providing the delegator's signature through an endpoint",
-                origin: TEST_ACCOUNTS.TRANSACTION.TRANSACTION_SENDER,
-                options: {
-                    delegatorUrl: 'https://example.com'
-                } satisfies SignTransactionOptions,
-                isDelegated: true,
-                expectedError: TransactionDelegationError
-            }
-        ]
-    },
-    testnet: {
-        correct: [
-            {
-                description: 'Should sign a transaction with delegation url',
-                origin: TEST_ACCOUNTS.TRANSACTION.TRANSACTION_SENDER,
-                options: {
-                    delegatorUrl: TESTNET_DELEGATE_URL
-                } satisfies SignTransactionOptions,
-                isDelegated: true,
-                expected: {
-                    body: {
-                        chainTag: 39,
-                        clauses: [
-                            {
-                                data: '0x01cb08c5000000000000000000000000000000000000000000000000000000000000007b',
-                                to: '0xb2c20a6de401003a671659b10629eb82ff254fb8',
-                                value: 0
-                            }
-                        ],
-                        dependsOn: null,
-                        expiration: 32,
-                        gas: 21464,
-                        gasPriceCoef: 0,
-                        reserved: {
-                            features: 1
-                        }
-                    }
-                }
-            }
-        ]
-    }
-};
-
 export {
     transactionNonces,
     waitForTransactionTestCases,
@@ -430,6 +300,5 @@ export {
     transferTransactionBodyValueAsNumber,
     expectedReceipt,
     transfer1VTHOClause,
-    buildTransactionBodyClausesTestCases,
-    signTransactionTestCases
+    buildTransactionBodyClausesTestCases
 };
