+++ packages/network/tests/thor-client/transactions/transactions.testnet.test.ts
@@ -1,28 +1,7 @@
 import { afterEach, beforeEach, describe, expect, test } from '@jest/globals';
-import {
-    buildTransactionBodyClausesTestCases,
-    signTransactionTestCases
-} from './fixture';
-import {
-    TESTING_CONTRACT_ABI,
-    TESTING_CONTRACT_ADDRESS,
-    testnetUrl,
-    THOR_SOLO_ACCOUNTS_BASE_WALLET
-} from '../../fixture';
-import {
-    addressUtils,
-    clauseBuilder,
-    coder,
-    type FunctionFragment,
-    TransactionHandler
-} from '@vechain/sdk-core';
-import {
-    ProviderInternalBaseWallet,
-    signerUtils,
-    ThorClient,
-    VechainBaseSigner,
-    VechainProvider
-} from '../../../src';
+import { buildTransactionBodyClausesTestCases } from './fixture';
+import { testnetUrl, THOR_SOLO_ACCOUNTS_BASE_WALLET } from '../../fixture';
+import { ThorClient, VechainProvider } from '../../../src';
 
 /**
  * Transactions module tests suite.
@@ -94,82 +73,6 @@ describe('Transactions module Testnet tests suite', () => {
                         expected.testnet.reserved
                     );
                     expect(txBody.chainTag).toBe(expected.testnet.chainTag);
-                });
-            }
-        );
-    });
-
-    /**
-     * Test suite for signTransaction method
-     */
-    describe('signTransactionTestCases', () => {
-        signTransactionTestCases.testnet.correct.forEach(
-            ({ description, origin, options, isDelegated, expected }) => {
-                test(description, async () => {
-                    const thorClient = ThorClient.fromUrl(testnetUrl);
-
-                    const sampleClause = clauseBuilder.functionInteraction(
-                        TESTING_CONTRACT_ADDRESS,
-                        coder
-                            .createInterface(TESTING_CONTRACT_ABI)
-                            .getFunction(
-                                'setStateVariable'
-                            ) as FunctionFragment,
-                        [123]
-                    );
-
-                    const gasResult = await thorClient.gas.estimateGas(
-                        [sampleClause],
-                        origin.address
-                    );
-
-                    const txBody =
-                        await thorClient.transactions.buildTransactionBody(
-                            [sampleClause],
-                            gasResult.totalGas,
-                            {
-                                isDelegated
-                            }
-                        );
-
-                    // Get the signer and sign the transaction
-                    const signer = new VechainBaseSigner(
-                        Buffer.from(origin.privateKey, 'hex'),
-                        new VechainProvider(
-                            thorClient,
-                            new ProviderInternalBaseWallet([], {
-                                delegator: options
-                            }),
-                            isDelegated
-                        )
-                    );
-
-                    const signedRawTx = isDelegated
-                        ? await signer.signTransactionWithDelegator(
-                              signerUtils.transactionBodyToTransactionRequestInput(
-                                  txBody,
-                                  origin.address
-                              )
-                          )
-                        : await signer.signTransaction(
-                              signerUtils.transactionBodyToTransactionRequestInput(
-                                  txBody,
-                                  origin.address
-                              )
-                          );
-                    const signedTx = TransactionHandler.decode(
-                        Buffer.from(signedRawTx.slice(2), 'hex'),
-                        true
-                    );
-
-                    expect(signedTx).toBeDefined();
-                    expect(signedTx.body).toMatchObject(expected.body);
-                    expect(signedTx.origin).toBe(
-                        addressUtils.toERC55Checksum(origin.address)
-                    );
-                    expect(signedTx.isDelegated).toBe(isDelegated);
-                    expect(signedTx.isSigned).toBe(true);
-                    expect(signedTx.signature).toBeDefined();
                 });
             }
         );
